+++ yellowbox/extras/redis/__init__.py
@@ -1,0 +1,4 @@
+from yellowbox.extras.redis.base import REDIS_DEFAULT_PORT, DEFAULT_RDB_PATH, BaseRedisService
+from yellowbox.extras.redis.redis import RedisService, append_state
+
+__all__ = ['REDIS_DEFAULT_PORT', 'DEFAULT_RDB_PATH', 'BaseRedisService', 'RedisService', 'append_state']